Preserve OpenAI Hermes codecs across module split
